feat(yard): add types and some sanity
