+++ documentation-site/pages/components/button.mdx
@@ -29,7 +29,7 @@ export default Layout;
 
 # Button
 
-<Yard />
+<Yard componentName="Button" />
 
 Buttons provide cues for actions and events. These fundamental components allow users to process actions or navigate an experience.
